--- tomcat/src/main/java/org/apache/coyote/http11/exception/unauthorised/InvalidSessionException.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.exception.unauthorised;
-
-public class InvalidSessionException extends UnAuthorisedException {
-
-    public InvalidSessionException() {
-        super("해당 세션이 존재하지 않습니다.");
-    }
-}
