--- tomcat/src/main/java/org/apache/coyote/http11/exception/unauthorised/LoginFailException.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.exception.unauthorised;
-
-public class LoginFailException extends UnAuthorisedException {
-
-    public LoginFailException() {
-        super("로그인에 실패했습니다.");
-    }
-}
