--- tomcat/src/main/java/org/apache/coyote/http11/exception/unauthorised/UnAuthorisedException.java
@@ -1,8 +1,0 @@
-package org.apache.coyote.http11.exception.unauthorised;
-
-public abstract class UnAuthorisedException extends RuntimeException {
-
-    public UnAuthorisedException(final String message) {
-        super(message);
-    }
-}
