+++ tomcat/src/main/java/org/apache/coyote/http11/exception/unauthorized/UnAuthorizedException.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.http11.exception.unauthorized;
+
+public abstract class UnAuthorizedException extends RuntimeException {
+
+    public UnAuthorizedException(final String message) {
+        super(message);
+    }
+}
